+++ packages/design-system/src/components/InlineEditing/InlineEditing.test.tsx
@@ -31,4 +31,20 @@ describe('InlineEditing', () => {
 		const results = await axe(document.body);
 		expect(results).toHaveNoViolations();
 	});
+
+	it('should render with data-testid prefix', async () => {
+		render(
+			<main>
+				<InlineEditing
+					label="Edit the value"
+					placeholder="What is your Lorem Ipsum?"
+					defaultValue="Lorem Ipsum"
+					onEdit={jest.fn()}
+					data-testid="my-prefix"
+				/>
+			</main>,
+		);
+
+		expect(screen.getByTestId('my-prefix.inlineediting.button.edit')).toBeInTheDocument();
+	});
 });