+++ packages/faceted-search/src/components/AdvancedSearch/AdvancedSearch.component.js
@@ -1,20 +1,18 @@
 /* eslint-disable jsx-a11y/no-autofocus */
 import { useState } from 'react';
 
+import classnames from 'classnames';
 import PropTypes from 'prop-types';
 
 import { ButtonIcon, Icon } from '@talend/design-system';
-import { getTheme } from '@talend/react-components/lib/theme';
 
 import { USAGE_TRACKING_TAGS } from '../../constants';
 import { useFacetedSearchContext } from '../context/facetedSearch.context';
 
-import theme from './AdvancedSearch.module.scss';
-
-const css = getTheme(theme);
+import styles from './AdvancedSearch.module.scss';
 
 const AdvancedSearchError = ({ id, label }) => (
-	<p aria-live="assertive" className={css('adv-search-error')} id={`${id}-error`} role="status">
+	<p aria-live="assertive" className={styles['adv-search-error']} id={`${id}-error`} role="status">
 		{label}
 	</p>
 );
@@ -69,9 +67,9 @@ export function AdvancedSearch({
 	};
 	const advSearchId = `${id}-adv-search`;
 	return (
-		<div id={advSearchId} className={css('adv-search')}>
+		<div id={advSearchId} className={styles['adv-search']}>
 			<form id={`${advSearchId}-form`} role="search" onSubmit={formSubmit}>
-				<Icon name="talend-filter" size="M" className={css('adv-search-filter-icon')} />
+				<Icon name="talend-filter" size="M" className={styles['adv-search-filter-icon']} />
 				<input
 					id={`${id}-form`}
 					name="advanced-search-faceted"
@@ -79,7 +77,7 @@ export function AdvancedSearch({
 					value={query}
 					placeholder={placeholder || t('ADV_SEARCH_FACETED_PLACEHOLDER', 'Enter your query')}
 					autoComplete="off"
-					className={css('adv-search-input', { 'has-error': error })}
+					className={classnames(styles['adv-search-input'], { 'has-error': error })}
 					aria-label={placeholder || t('ADV_SEARCH_FACETED_ARIA', 'Advanced Faceted Search')}
 					autoFocus
 					role="searchbox"
@@ -87,7 +85,7 @@ export function AdvancedSearch({
 					onChange={onChangeHandler}
 				/>
 
-				<div className={css('adv-search-buttons')}>
+				<div className={styles['adv-search-buttons']}>
 					<ButtonIcon
 						name="action-cancel-title"
 						icon="cross-filled"