+++ packages/faceted-search/src/dictionary/badge.dictionary.js
@@ -4,7 +4,6 @@ import { BadgeMenu } from '../components/Badges/BadgeMenu/BadgeMenu.component';
 import { BadgeNumber } from '../components/Badges/BadgeNumber/BadgeNumber.component';
 import { BadgePeriod } from '../components/Badges/BadgePeriod/BadgePeriod.component';
 import { BadgeSlider } from '../components/Badges/BadgeSlider/BadgeSlider.component';
-import { BadgeTags } from '../components/Badges/BadgeTags/BadgeTags.component';
 import { BadgeText } from '../components/Badges/BadgeText/BadgeText.component';
 
 const standardBadgeTypeNames = {
@@ -23,7 +22,7 @@ const standardBadges = {
 	[standardBadgeTypeNames.checkbox]: BadgeCheckboxes,
 	[standardBadgeTypeNames.number]: BadgeNumber,
 	[standardBadgeTypeNames.slider]: BadgeSlider,
-	[standardBadgeTypeNames.tags]: BadgeTags,
+	[standardBadgeTypeNames.tags]: BadgeCheckboxes,
 	[standardBadgeTypeNames.date]: BadgeDate,
 	[standardBadgeTypeNames.menu]: BadgeMenu,
 	[standardBadgeTypeNames.period]: BadgePeriod,